% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_rmd_contrasts.R
\name{bioinfo_rmd_contrasts}
\alias{bioinfo_rmd_contrasts}
\title{Write Bioinfo RMD}
\usage{
bioinfo_rmd_contrasts(filename, yaml.title, local.path = NULL,
  data.type = "Gene expression", input.files, data.logged = TRUE,
  data.nas = TRUE, min.npergrp = 3, grp.var = "grp", covars = NULL,
  aw.model = paste0("~0+", grp.var), use_aw = TRUE, use_trend = FALSE,
  contr.v = NULL, limma.model = NULL, row.type = "gene")
}
\arguments{
\item{filename}{Filename of RMD.}

\item{yaml.title}{Title of RMD.}

\item{data.type}{Description of expression and phenotype data.}

\item{input.files}{Character vector of input files. Matrix, pheno, and (optionally) annot.}

\item{data.logged}{Logical indicating if data has been log2-transformed.}

\item{data.nas}{Logical indicating if data has NAs.}

\item{min.npergrp}{Minimum sample size per group.}

\item{grp.var}{Variable name in \code{pheno} for group.}

\item{covars}{Variable name in \code{pheno} for covariates.}

\item{aw.model}{Model formula for \code{arrayWeights}.}

\item{use_aw}{Logical indicating if array weights should be used.}

\item{use_trend}{Logical indicating if \code{limma trend} should be used.}

\item{contr.v}{Text defining named vector of contrasts. If \code{NULL}, all pairwise contrasts are tested.}

\item{limma.model}{Model formula for \code{limma}, if want a design matrix.}

\item{row.type}{Character in filename for features.}

\item{path}{Path of RMD.}
}
\description{
Write Bioinformatics RMD workflow.
}
\details{
If need to remove a sample, rerun with new \code{input.files}.
}
