% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_chunk.R
\name{boxplot_chunk}
\alias{boxplot_chunk}
\title{Boxplot chunk}
\usage{
boxplot_chunk(elist = FALSE)
}
\arguments{
\item{elst}{Logical indicating if expression object an EList.}
}
\description{
Boxplot chunk.
}
