% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_txt.R
\name{data_txt}
\alias{data_txt}
\title{Write data description text}
\usage{
data_txt(input.files = NULL, path)
}
\arguments{
\item{input.files}{Character vector of input files.}

\item{path}{The path to the input files.}
}
\description{
Write data description text.
}
