% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_filt_voom_chunk.R
\name{feat_filt_voom_chunk}
\alias{feat_filt_voom_chunk}
\title{Feature filter chunk}
\usage{
feat_filt_voom_chunk(grp.var, row.type = "row")
}
\arguments{
\item{grp.var}{Variable name in \code{pheno} for group.}

\item{row.type}{Character in filename for features.}
}
\description{
Feature filter chunk.
}
\details{
A feature must be present at least 80% of \code{min.npergrp}.
}
