% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plots_chunk.R
\name{feature_plots_chunk}
\alias{feature_plots_chunk}
\title{Feature plots chunk}
\usage{
feature_plots_chunk(grp.var = "grp", path, use_annot = TRUE)
}
\arguments{
\item{grp.var}{Variable name in \code{pheno} for group.}

\item{path}{Path of RMD.}

\item{use_annot}{Should \code{annot} be used.}
}
\description{
Feature plots chunk.
}
