% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanvar_trend_chunk.R
\name{meanvar_trend_chunk}
\alias{meanvar_trend_chunk}
\title{Meanvar trend chunk}
\usage{
meanvar_trend_chunk(voom.model, use_trend = FALSE)
}
\arguments{
\item{voom.model}{Model formula for \code{arrayWeights}.}

\item{use_trend}{Logical indicating if \code{limma trend} should be used.}
}
\description{
Meanvar trend chunk.
}
