% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_chunk.R
\name{norm_chunk}
\alias{norm_chunk}
\title{Normalization chunk}
\usage{
norm_chunk(proj.nm, path)
}
\arguments{
\item{proj.nm}{Project name.}

\item{path}{Path of RMD.}
}
\description{
Normalization chunk.
}
