% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_chunk.R
\name{pca_chunk}
\alias{pca_chunk}
\title{PCA chunk}
\usage{
pca_chunk(grp.var, covars = NULL)
}
\arguments{
\item{grp.var}{Variable name in \code{pheno} for group.}

\item{covars}{Variable name in \code{pheno} for covariates.}
}
\description{
PCA chunk.
}
