% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_chunk.R
\name{read_data_chunk}
\alias{read_data_chunk}
\title{Read chunk}
\usage{
read_data_chunk(input.files, data.logged)
}
\arguments{
\item{input.files}{Character vector of input files. Should be data matrix, pheno, then optionally annot.}

\item{data.logged}{Logical indicating if data has been log2-transformed.}
}
\description{
Read chunk
}
