% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_links.R
\name{rmd_links}
\alias{rmd_links}
\title{Create RMD links}
\usage{
rmd_links(filenames, path = NULL)
}
\arguments{
\item{filenames}{Filenames to link to, possibly named.}

\item{path}{Path to prefix.}
}
\description{
Create RMD links.
}
