% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roast_contrasts_chunk.R
\name{roast_contrasts_chunk}
\alias{roast_contrasts_chunk}
\title{Roast contrasts chunk}
\usage{
roast_contrasts_chunk(
  grp.var,
  path,
  pdb.files = "cp = 'c2.cp.v6.0'",
  use_des = FALSE,
  use_aw = TRUE,
  use_trend = FALSE,
  elst = FALSE
)
}
\arguments{
\item{grp.var}{Variable name in \code{pheno} for group.}

\item{path}{Path of RMD.}

\item{pdb.files}{Text defining pathway database files.}

\item{use_des}{Logical indicating if design (from limma chunk) should be used.}

\item{use_aw}{Logical indicating if array weights should be used.}

\item{use_trend}{Logical indicating if \code{limma trend} should be used.}

\item{elst}{Logical indicating if expression object an EList.}
}
\description{
Roast contrasts chunk.
}
