% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_chunk.R
\name{setup_chunk}
\alias{setup_chunk}
\title{Write setup chunk}
\usage{
setup_chunk(path, link2wd = TRUE)
}
\arguments{
\item{path}{Local path of working directory}

\item{link2wd}{Boolean; should links be to working directory on Joslin network? Otherwise, links are to local folder}
}
\description{
Write setup chunk.
}
