% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bioinfo_rmd.R
\name{write_bioinfo_rmd}
\alias{write_bioinfo_rmd}
\title{Write Bioinfo RMD}
\usage{
write_bioinfo_rmd(filename, yaml.title, local.path = NULL,
  data.type = "Gene expression", input.files, data.logged = TRUE,
  data.nas = TRUE, min.npergrp = 3, grp.var = "grp",
  aw.model = paste0("~0+", grp.var), covars = NULL)
}
\arguments{
\item{filename}{Filename of RMD.}

\item{yaml.title}{Title of RMD.}

\item{data.type}{Description of expression and phenotype data.}

\item{input.files}{Character vector of input files.}

\item{data.logged}{Logical indicating if data has been log2-transformed.}

\item{data.nas}{Logical indicating if data has NAs.}

\item{min.npergrp}{Minimum sample size per group.}

\item{grp.var}{Variable name in \code{pheno} for group.}

\item{aw.model}{Model formula for \code{arrayWeights}.}

\item{covars}{Variable name in \code{pheno} for covariates.}

\item{path}{Path of RMD.}
}
\description{
Write Bioinformatics RMD workflow.
}
