% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_blocks.R
\name{write_blocks}
\alias{write_blocks}
\title{Write blocks of text and code}
\usage{
write_blocks(filename, blocks)
}
\arguments{
\item{filename}{Filename to write to. \code{Rmd} extension is added.}

\item{blocks}{List of blocks to write.}
}
\description{
Write blocks of text and code.
}
