% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data_chunk.R
\name{write_data_chunk}
\alias{write_data_chunk}
\title{Write data matrix}
\usage{
write_data_chunk(proj.nm, use_annot = TRUE, row.type = "gene")
}
\arguments{
\item{proj.nm}{Name of project to paste into filenames.}

\item{use_annot}{Should \code{annot} be used.}

\item{row.type}{Character in filename for features.}
}
\description{
Write imputed, filtered, normalized data for plotting.
}
