% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_setup_chunk.R
\name{write_setup_chunk}
\alias{write_setup_chunk}
\title{Write setup chunk}
\usage{
write_setup_chunk(path = NULL)
}
\arguments{
\item{path}{Path of working directory.}
}
\description{
Write setup chunk.
}
