% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yaml_header.R
\name{write_yaml_header}
\alias{write_yaml_header}
\title{Write YAML header}
\usage{
write_yaml_header(title)
}
\arguments{
\item{title}{Title of project}
}
\description{
Write YAML header for project.
}
