% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_rmd_contrasts.R
\name{bioinfo_rmd_contrasts}
\alias{bioinfo_rmd_contrasts}
\title{Write Bioinfo RMD}
\usage{
bioinfo_rmd_contrasts(
  filename,
  local.path = NULL,
  data.desc = "Gene expression",
  input.files,
  data.logged = TRUE,
  data.nas = TRUE,
  grp.var = "grp",
  covars = NULL,
  aw.model = paste0("~0+", grp.var),
  use_aw = TRUE,
  use_trend = FALSE,
  contr.v,
  limma.model = NULL,
  row.type = "gene",
  gmt_abbrev = c("reactome", "tft"),
  gmt_prefix = c("c2.cp.reactome", "c3.tft.gtrd")
)
}
\arguments{
\item{filename}{Filename of RMD.}

\item{local.path}{Path of RMD.}

\item{data.desc}{Description of data sources for \code{input.files}.}

\item{input.files}{Character vector of input files. Matrix, pheno, and (optionally) annot.}

\item{data.logged}{Logical indicating if data has been log2-transformed.}

\item{data.nas}{Logical indicating if data has NAs.}

\item{grp.var}{Variable name in \code{pheno} for group.}

\item{covars}{Variable name in \code{pheno} for covariates.}

\item{aw.model}{Model for `limma::arrayWeights`.}

\item{use_aw}{Logical indicating if array weights should be used.}

\item{use_trend}{Logical indicating if \code{limma trend} should be used.}

\item{contr.v}{Named vector of contrasts.}

\item{limma.model}{Model formula for \code{limma}, if want a design matrix.}

\item{row.type}{Character in filename for features.}

\item{gmt_abbrev}{Character vector of abbreviation of pathway name, e.g. "reactome" or "tft".}

\item{gmt_prefix}{Character vector of GMT file(s) prefixes.}
}
\description{
Write Bioinformatics RMD workflow.
}
\details{
If need to remove a sample, rerun with new \code{input.files}.
}
\examples{
\donttest{
bioinfo_rmd_contrasts(filename="new_analysis", input.files = c("counts.csv", "pheno.csv"),
contr.v='c(treat="treat-control")')
}
}
