% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_txt.R
\name{data_txt}
\alias{data_txt}
\title{Write data description text}
\usage{
data_txt(data.type = "Gene expression", input.files = NULL, path)
}
\arguments{
\item{data.type}{Description of biomaterial and condition.}

\item{input.files}{Character vector of input files.}
}
\description{
Write data description text.
}
