% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_contrasts_chunk.R
\name{limma_contrasts_chunk}
\alias{limma_contrasts_chunk}
\title{Limma contrasts chunk}
\usage{
limma_contrasts_chunk(grp.var, contr.v, path, limma.model = NULL,
  use_aw = TRUE, use_trend = FALSE, use_annot = TRUE, row.type = "gene")
}
\arguments{
\item{grp.var}{Variable name in \code{pheno} for group.}

\item{contr.v}{Text defining named vector of contrasts. If \code{NULL}, all pairwise contrasts are tested.}

\item{path}{Path of RMD.}

\item{limma.model}{Model formula for \code{arrayWeights}.}

\item{use_aw}{Logical indicating if array weights should be used.}

\item{use_trend}{Logical indicating if \code{limma trend} should be used.}

\item{use_annot}{Should \code{annot} be used.}

\item{row.type}{Character in filename for features.}
}
\description{
Limma contrasts chunk.
}
