% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_chunk.R
\name{norm_chunk}
\alias{norm_chunk}
\title{Normalization chunk}
\usage{
norm_chunk(input.files, path)
}
\arguments{
\item{input.files}{Character vector of input files. Should be data.matrix, pheno, then optionally annot.}

\item{path}{Path of RMD.}
}
\description{
Normalization chunk.
}
