% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_voom_chunk.R
\name{norm_voom_chunk}
\alias{norm_voom_chunk}
\title{Normalization chunk}
\usage{
norm_voom_chunk(proj.nm, voom.model, path, use_aw)
}
\arguments{
\item{proj.nm}{Name of project.}

\item{voom.model}{Model for design matrix.}

\item{path}{Path of RMD.}

\item{use_aw}{Logical indicating if array weights should be used.}
}
\description{
Normalization chunk.
}
