% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_chunk.R
\name{pca_chunk}
\alias{pca_chunk}
\title{PCA chunk}
\usage{
pca_chunk(grp.var, proj.nm, covars = NULL, elst = FALSE)
}
\arguments{
\item{grp.var}{Variable name in \code{pheno} for group.}

\item{proj.nm}{Name of project to paste into filenames.}

\item{covars}{Variable name in \code{pheno} for covariates.}

\item{elst}{Logical indicating if expression object an EList.}
}
\description{
PCA chunk.
}
