% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roast_contrasts_chunk.R
\name{roast_contrasts_chunk}
\alias{roast_contrasts_chunk}
\title{Roast contrasts chunk}
\usage{
roast_contrasts_chunk(
  grp.var,
  path,
  gmt_abbrev = c("reactome", "tft", "mir"),
  gmt_prefix = c("c2.cp.reactome", "c3.tft.gtrd", "c3.mir.mirdb"),
  use_des = FALSE,
  use_aw = TRUE,
  use_trend = FALSE,
  elst = FALSE
)
}
\arguments{
\item{grp.var}{Variable name in \code{pheno} for group.}

\item{path}{Path of RMD.}

\item{gmt_abbrev}{Character vector of abbreviation of pathway name, e.g. "reactome" or "tft".}

\item{gmt_prefix}{Character vector of GMT file(s) prefixes.}

\item{use_des}{Logical indicating if design (from limma chunk) should be used.}

\item{use_aw}{Logical indicating if array weights should be used.}

\item{use_trend}{Logical indicating if \code{limma trend} should be used.}

\item{elst}{Logical indicating if expression object an EList.}
}
\description{
Roast contrasts chunk.
}
