% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bioinfo_rmd.R
\name{write_bioinfo_rmd}
\alias{write_bioinfo_rmd}
\title{Write Bioinfo RMD}
\usage{
write_bioinfo_rmd(filename, title, path = NULL)
}
\arguments{
\item{filename}{Filename of RMD.}

\item{title}{Title of RMD.}

\item{path}{Path of RMD.}
}
\description{
Write Bioinformatics RMD workflow
}
